[chore] #114 DateService API

+++ app/src/main/java/org/sopt/dateroad/data/dataremote/datasource/DateRemoteDataSource.kt
@@ -5,8 +5,11 @@ import org.sopt.dateroad.data.dataremote.model.response.ResponseDatesDto
 import org.sopt.dateroad.data.dataremote.model.response.ResponseNearestDateDto
 
 interface DateRemoteDataSource {
-    suspend fun deleteDate(dateId: Long)
-    suspend fun getDateDetail(dateId: Long): ResponseDateDetailDto
+    suspend fun deleteDate(dateId: Int)
+
+    suspend fun getDateDetail(dateId: Int): ResponseDateDetailDto
+
     suspend fun getDates(time: String): ResponseDatesDto
+
     suspend fun getNearestDate(): ResponseNearestDateDto
 }
